Terasort_with_virtual_memory_and_paging
Terasort of 128GB and 1TB datasets: 
Used C to have more control over memory allocation and de-allocation and used the merge sort which was the best fit for this task and completely on AWS Ubuntu instances.
Worked in 3 modes. 
1. Hadoop 
2. Spark 
3. MPI (Message Passing Interface) 
which confirmed that Hadoop was the slowest among the 3 with MPI giving the highest throughput but Spark being the easiest to implement with very close results relative to Spark. 
Used Virtual memory and paging on AWS ec2 i3.4xlarge nodes as the required memory was at least 9x the available memory.